Proposal UI done
