+++ src/containers/Login/Login.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useAppSelector, useAppDispatch } from '../../store/configureStore';
-import { getEmail, getPassword } from '../../features/login/loginSlice';
+import { getEmail, getPassword } from '../../slice/login/loginSlice';
 import { Input } from '../../components/Input/Input'
 import About from '../../components/About/About';
