--- src/features/proposal/proposalType.ts
@@ -1,37 +1,0 @@
-export type proposalDataStates = {
-    proposalType: string,
-    name: string,
-    email: string,
-    startTime: {
-        year: string,
-        month: string,
-        day: string,
-        hour: string,
-        minute: string,
-    },
-    endTime: {
-        year: string,
-        month: string,
-        day: string,
-        hour: string,
-        minute: string,
-    },
-    targetAmount: string,
-    category: string,
-    planName: string,
-    planSummary: string,
-    coverPhoto: string,
-    planIntroduction: string,
-    rewardChoiceListType: {
-        theFirstOfReward: {
-            ammount: string,
-            introduction: string,
-        },
-        theSecondOfReward: {
-            ammonut: string,
-            introduction: string,
-        }
-    },
-    webLink: string,
-    VideoLink: string,
-};
